# modules/apps/headless/headless-admin-workflow/headless-admin-workflow-impl/rest-openapi.yaml
components:
    schemas:
...
        WorkflowDefinition:
            properties:
                active:
                    type: boolean
                content:
                    type: string
                dateModified:
                    format: date-time
                    readOnly: true
                    type: string
                description:
                    readOnly: true
                    type: string
                name:
                    type: string
                title:
                    type: string
                version:
                    type: string
            type: object
...
info:
    description:
        "A Java client JAR is available for use with the group ID 'com.liferay', artifact ID
        'com.liferay.headless.admin.workflow.client', and version '2.0.2'."
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: "Headless Admin Workflow"
    version: v1.0
openapi: 3.0.1
paths:
    "/workflow-definitions":
        get:
            operationId: getWorkflowDefinitionsPage
            parameters:
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/WorkflowDefinition"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/WorkflowDefinition"
                                type: array
                    description:
                        ""
            tags: ["WorkflowDefinition"]
    "/workflow-definitions/by-name/{name}":
        get:
            operationId: getWorkflowDefinitionByName
            parameters:
                - in: path
                  name: name
                  required: true
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/WorkflowDefinition"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/WorkflowDefinition"
                    description:
                        ""
            tags: ["WorkflowDefinition"]
    "/workflow-definitions/deploy":
        post:
            operationId: postWorkflowDefinitionDeploy
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/WorkflowDefinition"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/WorkflowDefinition"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/WorkflowDefinition"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/WorkflowDefinition"
                    description:
                        ""
            tags: ["WorkflowDefinition"]
    "/workflow-definitions/save":
        post:
            operationId: postWorkflowDefinitionSave
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/WorkflowDefinition"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/WorkflowDefinition"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/WorkflowDefinition"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/WorkflowDefinition"
                    description:
                        ""
            tags: ["WorkflowDefinition"]
    "/workflow-definitions/undeploy":
        delete:
            operationId: deleteWorkflowDefinitionUndeploy
            parameters:
                - in: query
                  name: name
                  required: true
                  schema:
                      type: string
                - in: query
                  name: version
                  required: true
                  schema:
                      type: string
            responses:
                204:
                    content:
                        application/json: {}
                        application/xml: {}
                    description:
                        ""
            tags: ["WorkflowDefinition"]
    "/workflow-definitions/update-active":
        post:
            operationId: postWorkflowDefinitionUpdateActive
            parameters:
                - in: query
                  name: active
                  required: true
                  schema:
                      type: boolean
                - in: query
                  name: name
                  required: true
                  schema:
                      type: string
                - in: query
                  name: version
                  required: true
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/WorkflowDefinition"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/WorkflowDefinition"
                    description:
                        ""
            tags: ["WorkflowDefinition"]
...